requirements.txt
# requirements file
# sandbox had python 3.10

together==1.2.0
python-dotenv~=1.0.1
pydantic==2.8.2
gradio==4.44.1